requirements.txt
python-dotenv~=0.19.0
aiogram~=2.19
trio~=0.20.0